+++ src/Symfony/Bridge/PsrHttpMessage/Tests/Fixtures/App/Kernel.php
@@ -52,6 +52,8 @@ class Kernel extends SymfonyKernel
             'test' => true,
             'annotations' => false,
             'http_method_override' => false,
+            'handle_all_throwables' => true,
+            'php_errors' => ['log' => true],
         ]);
 
         $container->services()