+++ src/Symfony/Bundle/FrameworkBundle/Tests/Command/AboutCommand/Fixture/TestAppKernel.php
@@ -36,6 +36,8 @@ class TestAppKernel extends Kernel
             $container->loadFromExtension('framework', [
                 'annotations' => false,
                 'http_method_override' => false,
+                'handle_all_throwables' => true,
+                'php_errors' => ['log' => true],
             ]);
         });
     }