+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer_with_dsn.php
@@ -6,6 +6,8 @@ return static function (ContainerConfigurator $container) {
     $container->extension('framework', [
         'annotations' => false,
         'http_method_override' => false,
+    'handle_all_throwables' => true,
+    'php_errors' => ['log' => true],
         'mailer' => [
             'dsn' => 'smtp://example.com',
             'envelope' => [