+++ src/Symfony/Bundle/FrameworkBundle/Tests/Kernel/MicroKernelTraitTest.php
@@ -122,6 +122,8 @@ class MicroKernelTraitTest extends TestCase
                 $c->extension('framework', [
                     'annotations' => false,
                     'http_method_override' => false,
+                    'handle_all_throwables' => true,
+                    'php_errors' => ['log' => true],
                     'router' => ['utf8' => true],
                 ]);
                 $c->services()->set('logger', NullLogger::class);