+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Compiler/AddSessionDomainConstraintPassTest.php
@@ -26,7 +26,7 @@ class AddSessionDomainConstraintPassTest extends TestCase
 {
     public function testSessionCookie()
     {
-        $container = $this->createContainer(['cookie_domain' => '.symfony.com.', 'cookie_secure' => true]);
+        $container = $this->createContainer(['cookie_domain' => '.symfony.com.', 'cookie_secure' => true, 'cookie_samesite' => 'lax']);
 
         $utils = $container->get('security.http_utils');
         $request = Request::create('/', 'get');
@@ -41,7 +41,7 @@ class AddSessionDomainConstraintPassTest extends TestCase
 
     public function testSessionNoDomain()
     {
-        $container = $this->createContainer(['cookie_secure' => true]);
+        $container = $this->createContainer(['cookie_secure' => true, 'cookie_samesite' => 'lax']);
 
         $utils = $container->get('security.http_utils');
         $request = Request::create('/', 'get');
@@ -56,7 +56,7 @@ class AddSessionDomainConstraintPassTest extends TestCase
 
     public function testSessionNoSecure()
     {
-        $container = $this->createContainer(['cookie_domain' => '.symfony.com.']);
+        $container = $this->createContainer(['cookie_domain' => '.symfony.com.', 'cookie_samesite' => 'lax']);
 
         $utils = $container->get('security.http_utils');
         $request = Request::create('/', 'get');
@@ -102,7 +102,7 @@ class AddSessionDomainConstraintPassTest extends TestCase
 
     public function testSessionAutoSecure()
     {
-        $container = $this->createContainer(['cookie_domain' => '.symfony.com.', 'cookie_secure' => 'auto']);
+        $container = $this->createContainer(['cookie_domain' => '.symfony.com.', 'cookie_secure' => 'auto', 'cookie_samesite' => 'lax']);
 
         $utils = $container->get('security.http_utils');
         $request = Request::create('/', 'get');
@@ -145,7 +145,7 @@ class AddSessionDomainConstraintPassTest extends TestCase
         ];
 
         $ext = new FrameworkExtension();
-        $ext->load(['framework' => ['annotations' => false, 'http_method_override' => false, 'csrf_protection' => false, 'router' => ['resource' => 'dummy', 'utf8' => true]]], $container);
+        $ext->load(['framework' => ['annotations' => false, 'http_method_override' => false, 'handle_all_throwables' => true, 'php_errors' => ['log' => true], 'csrf_protection' => false, 'router' => ['resource' => 'dummy', 'utf8' => true]]], $container);
 
         $ext = new SecurityExtension();
         $ext->load($config, $container);