+++ src/Symfony/Component/AssetMapper/Tests/fixtures/AssetMapperTestAppKernel.php
@@ -38,6 +38,8 @@ class AssetMapperTestAppKernel extends Kernel
             $container->loadFromExtension('framework', [
                 'annotations' => false,
                 'http_method_override' => false,
+                'handle_all_throwables' => true,
+                'php_errors' => ['log' => true],
                 'http_client' => true,
                 'assets' => null,
                 'asset_mapper' => [